Fix infinite loop in AccountHandlingApi.getNewAddress(account)
