+++ testkit-core/.jvm/src/main/resources/logback-test.xml
@@ -29,7 +29,7 @@
 
     <logger name="org.bitcoins.chain" level="WARN"/>
 
-    <logger name="org.bitcoins.wallet" level="DEBUG"/>
+    <logger name="org.bitcoins.wallet" level="WARN"/>
 
     <logger name="org.bitcoins.dlc.wallet" level="WARN"/>
